+++ R/cardinal.R
@@ -1,10 +1,5 @@
 #' Convert numbers to cardinal character vectors (one, two, three)
 #'
-#' `nom_card()` and `cardinal()` produce cardinal numbers.
-#' `nom_numer()` and `numerator()` produce numerators.
-#' The results are equivalent for integers, but `nom_card()` and `cardinal()`
-#' support fractional components while `nom_numer()` and `numerator()` do not.
-#'
 #' @param x A numeric vector
 #' @param max_n A numeric vector.
 #'     When the absolute value of `x` is greater than `max_n`, `x` remains
@@ -12,16 +7,30 @@
 #'     If `max_n` is negative, no `x`s will be converted to words.
 #'     (This can be useful when `max_n` is passed by another function.)
 #'     Defaults to `Inf`, which converts all `x`s to words.
-#'     Default can be changed by setting `options("nombre.max_n")`.
+#'     Default can be changed with [set_config("nombre::max_n")][set_config()].
 #' @param negative A character vector to append to negative numbers.
 #'     Defaults to `"negative"`.
-#'     Default can be changed by setting `options("nombre.negative")`.
+#'     Default can be changed with
+#'     [set_config("nombre::negative")][set_config()].
+#' @param ... Additional arguments passed to [fracture::frac_mat()].
+#'     See details.
 #' @param numerator When `TRUE`, an error is produced if `x` has a decimal or
 #'     fractional component.
 #'     Defaults to `FALSE`.
-#' @param frac_args A list of arguments passed to [fracture::frac_mat()] when
-#'     converting decimals to fractions.
-#' @param ... Additional arguments of `numerator()` are passed to `cardinal()`
+#'
+#' @details # Fractions:
+#'
+#' Decimal components of `x` are automatically converted to fractions by
+#' [fracture::frac_mat()].
+#' Named arguments of `cardinal()` and `nom_card()` are passed to
+#' [fracture::frac_mat()] through [`...`][dots].
+#' Helpful arguments include:
+#' * `base_10 = TRUE`, which forces all fractions to use denominators that are
+#'   powers of ten
+#' * `common_denom = TRUE`, which forces all fractions to use the least common
+#'   denominator when `x` is a vector
+#' * `max_denom`, which sets the maximum allowable denominator.
+#'   By default, the maximum denominator is "ten-millionths".
 #'
 #' @return A character vector of the same length as `x`
 #' @family number names
@@ -30,9 +39,9 @@
 
 cardinal <- function(
   x,
-  max_n     = getOption("nombre.max_n", Inf),
-  negative  = getOption("nombre.negative", "negative"),
-  frac_args = list(),
+  max_n     = get_config("nombre::max_n", Inf),
+  negative  = get_config("nombre::negative", "negative"),
+  ...,
   numerator = FALSE
 ) {
   n <- length(x)
@@ -62,7 +71,7 @@ cardinal <- function(
   if (!numerator) {
     decimal                <- numeric(n)
     decimal[unmaxed]       <- x[unmaxed] %% 1
-    fraction[decimal != 0] <- convert_fraction(decimal[decimal != 0], frac_args)
+    fraction[decimal != 0] <- convert_fraction(decimal[decimal != 0], ...)
     x[unmaxed]             <- x[unmaxed] %/% 1
   }
 
@@ -101,13 +110,3 @@ cardinal <- function(
 #' @export
 
 nom_card <- cardinal
-
-#' @rdname cardinal
-#' @export
-
-numerator <- function(x, ...) {cardinal(x, ..., numerator = TRUE)}
-
-#' @rdname cardinal
-#' @export
-
-nom_numer <- numerator