+++ R/convert_fraction.R
@@ -1,7 +1,7 @@
-convert_fraction <- function(x, frac_args) {
+convert_fraction <- function(x, ...) {
   if (!length(x)) return(character(0))
 
-  x           <- do.call(fracture::frac_mat, c(list(x), frac_args))
+  x           <- fracture::frac_mat(x, ...)
   numerator   <- numerator(x[1, ])
   denominator <- denominator(x[2, ], x[1, ])